+++ symmetria/elements/_utils.py
@@ -1,34 +1,47 @@
-from collections import OrderedDict
+from typing import Dict, List, Self
+from dataclasses import field, dataclass
 
 
-def _pretty_print_table(title: str, body: OrderedDict[str, str]) -> str:
-    """Private method to print an ordered dictionary in a table format."""
-    max_length_keys = max([len(key) for key in body.keys()])
-    max_length_values = max([len(value) for value in body.values()])
-    max_length_body = max(max_length_values, max_length_keys)
+@dataclass(kw_only=True)
+class Table:
+    title: str
+    rows: Dict[str, str] = field(default_factory=dict)
 
-    if (len(title) + max_length_body) % 2 == 0:
-        length_table = len(title) + max_length_body + 20
-    else:
-        length_table = len(title) + max_length_body + 21
+    _max_length_row: int = 0
 
-    table = "+" + "-" * length_table + "+"
-    table += "\n"
-    table += "|" + "{:^{length}}".format(title, length=length_table) + "|"
-    table += "\n"
-    table += "+" + "-" * length_table + "+"
-    table += "\n"
+    def add(self, attribute: str, value: str) -> Self:
+        self._max_length_row = max(self._max_length_row, len(attribute) + len(value))
+        self.rows[attribute] = value
+        return self
 
-    for name, value in body.items():
-        table += _get_row(length_table // 2, a=name, b=value)
-        table += "\n"
-        table += "+" + "-" * (length_table // 2) + "+" + "-" * (length_table // 2 - 1) + "+"
-        table += "\n"
+    def build(self) -> str:
+        length_table = len(self.title) + self._max_length_row + 20 + (len(self.title) + self._max_length_row) % 2
 
-    # the -1 is to delete the lase "\n" from the table
-    return table[:-1]
+        table: List[str] = []
+        table.append(self._get_header(length_table=length_table))
 
+        for key, value in self.rows.items():
+            table.append(self._get_row(length_table // 2, a=key, b=value))
+            table.append(self._get_separation_for_rows(length_table=length_table))
 
-def _get_row(length: int, a: str, b: str) -> str:
-    """Return a row of the table."""
-    return "|" + "{:<{length}}".format(" " + a, length=length) + "|" + "{:^{length}}".format(b, length=length - 1) + "|"
+        return "\n".join(table)
+
+    def _get_header(self, length_table: int) -> str:
+        """Return the header of the table."""
+        return "\n".join(
+            [
+                "+" + "-" * length_table + "+",
+                f"|{self.title:^{length_table}}|",
+                "+" + "-" * length_table + "+",
+            ]
+        )
+
+    def _get_row(self, length: int, a: str, b: str) -> str:
+        """Return a row of the table."""
+        return (
+            "|" + "{:<{length}}".format(" " + a, length=length) + "|" + "{:^{length}}".format(b, length=length - 1) + "|"
+        )
+
+    def _get_separation_for_rows(self, length_table: int) -> str:
+        """Return a separation line for two rows."""
+        return "+" + "-" * (length_table // 2) + "+" + "-" * (length_table // 2 - 1) + "+"